pyproject.toml
[project]
name = "filmhouse-calendar-sync"
version = "0.1.0"
description = "Daily scraper for Filmhouse.sg screenings synced to Google Calendar"
requires-python = ">=3.11"
dependencies = [
    "scrapling[fetchers]>=0.4.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
